% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotateResults.R
\name{AnnotateResults}
\alias{AnnotateResults}
\title{Annotate \code{coMethDMR} Pipeline Results}
\usage{
AnnotateResults(lmmRes_df, arrayType = c("450k", "EPIC"))
}
\arguments{
\item{lmmRes_df}{A data frame returned by \code{\link{lmmTest}} or a parallel
implementation thereof. This data frame must contain the following
columns: \code{chrom} = the chromosome the region is on; \code{start} =
the region start point; \code{end} = the region end point; and
\code{regionType} = a character string marking from which region type the
region came from. See \strong{Details} for more information.}

\item{arrayType}{Type of array: 450k or EPIC}
}
\value{
a data frame with location of the genomic region's chromosome
   (\code{chrom}), start (\code{start}), and end (\code{end}); the number of
   CpGs in the region (\code{nCpGs}); results for testing association of
   methylation in individual CpGs with continuous phenotype (\code{Estimate},
   \code{StdErr}, \code{Stat}, and \code{pValue}); the type of region
   (\code{regionType}) and its relationship to known CpG islands
   (\code{Relation_to_UCSC_CpG_Island}); UCSC annotation information
   (\code{UCSC_RefGene_Group}, \code{UCSC_RefGene_Accession}, and
   \code{UCSC_RefGene_Name}); and a list of all of the probes in that region
   (\code{probes}).
}
\description{
Given a data frame with regions of the genome, add UCSC
   Reference Gene information, probe IDs, and relationship to known CpG
   islands.
}
\details{
The region types include \code{"NSHORE"}, \code{"NSHELF"},
   \code{"SSHORE"}, \code{"SSHELF"}, \code{"TSS1500"}, \code{"TSS200"},
   \code{"UTR5"}, \code{"EXON1"}, \code{"GENEBODY"}, \code{"UTR3"}, and
   \code{"ISLAND"}.
}
\examples{
   lmmResults_df <- data.frame(
     chrom = c("chr22", "chr22", "chr22", "chr22", "chr22"),
     start = c("39377790", "50987294", "19746156", "42470063", "43817258"),
     end   = c("39377930", "50987527", "19746368", "42470223", "43817384"),
     regionType = c("TSS1500", "EXON1", "ISLAND", "TSS200", "ISLAND"),
     stringsAsFactors = FALSE
   )

   AnnotateResults(
     lmmRes_df = lmmResults_df,
     arrayType = "450k"
   )

}
