% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_betaMatrix_ex1.R
\docType{data}
\name{betaMatrix_ex1}
\alias{betaMatrix_ex1}
\title{Alzheimer's Prefrontal Cortex (PFC) Methylation Data}
\format{
A data frame containing beta values for 4 CpGs in one CpG islands for 110 subjects.
Each column is a CpG, each row is a sample.
}
\source{
GEO accession: GSE59685
}
\usage{
betaMatrix_ex1
}
\description{
Subset of an Alzheimer's Disease methylation data set,
  with beta values for measured CpGs methylation levels.
}
\keyword{datasets}
