% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_betaMatrix_ex4.R
\docType{data}
\name{betaMatrix_ex4}
\alias{betaMatrix_ex4}
\title{Alzheimer's Prefrontal Cortex (PFC) Methylation Data}
\format{A subset of a data frame containing beta values for 52 CpGs in
  one CpG island in 110 subjects.}
\source{
GEO accession: GSE59685
}
\usage{
betaMatrix_ex4
}
\description{
Subset of an Alzheimer's methylation data set,
  with beta values for CpGs.
}
\keyword{datasets}
