% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_betaMatrixChr22_df.R
\docType{data}
\name{betaMatrixChr22_df}
\alias{betaMatrixChr22_df}
\title{Alzheimer's Prefrontal Cortex (PFC) Methylation Data}
\format{A subset of a data frame containing beta values for 8552 CpGs in
  Chr22 in 20 subjects.}
\source{
GEO accession: GSE59685
}
\usage{
betaMatrixChr22_df
}
\description{
Subset of an Alzheimer's methylation dataset,
  with beta values for CpGs.
}
\keyword{datasets}
