% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_betaMatrixChr22_df.R
\docType{data}
\name{betasChr22_df}
\alias{betasChr22_df}
\title{Prefrontal Cortex (PFC) Methylation Data from Alzheimer's Disease subjects}
\format{
A data frame containing beta values for 8552 CpGs in
  Chr22 for a subset of 20 subjects.
}
\source{
GEO accession: GSE59685
}
\usage{
betasChr22_df
}
\description{
Subset of an Alzheimer's methylation dataset, with beta values for CpGs.
}
\keyword{datasets}
