% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CloseByAllRegions.R
\name{CloseByAllRegions}
\alias{CloseByAllRegions}
\title{Extract clusters of close by CpGs from all genomic regions}
\usage{
CloseByAllRegions(genomicRegionType = c("TSS1500Ind", "TSS200Ind",
  "UTR5Ind", "EXON1Ind", "GENEBODYInd", "UTR3Ind", "ISLANDInd",
  "NSHOREInd", "SSHOREInd", "NSHELFInd", "SSHELFInd"),
  arrayType = c("450k", "EPIC"), maxGap, minCpGs, output = c("RDS",
  "GMT"), outputPath, ...)
}
\arguments{
\item{arrayType}{Type of array, 450k or EPIC}

\item{maxGap}{an integer, genomic locations within maxGap from each other
are placed into the same cluster}

\item{minCpGs}{an integer, minimum number of CpGs for resulting regions}

\item{output}{the output files can be saved as .gmt or .RDS}

\item{outputPath}{link to the folder where the filtered files will be saved}

\item{...}{}
}
\value{
a file with the close by regions
}
\description{
Extract clusters of close by CpGs from all genomic regions
}
