% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CloseBySingleRegion.R
\name{CloseBySingleRegion}
\alias{CloseBySingleRegion}
\title{Extract clusters of close by CpGs from a genomic region}
\usage{
CloseBySingleRegion(CpGs_char, arrayType = c("450k", "EPIC"),
  maxGap = 200, minCpGs = 3)
}
\arguments{
\item{CpGs_char}{a list of CpG ids}

\item{arrayType}{Type of array, 450k or EPIC}

\item{maxGap}{an integer, genomic locations within maxGap from each other
are placed into the same cluster}

\item{minCpGs}{an integer, minimum number of CpGs for resulting regions}
}
\value{
a list, each item is a list of CpG ids in the contiguous region
   (i.e. same cluster)
}
\description{
Extract clusters of close by CpGs from a genomic region
}
\examples{
   CpGs_char <- c("cg02505293", "cg03618257", "cg04421269", "cg17885402",
       "cg19890033", "cg20566587", "cg27505880")
   CloseBySingleRegion(CpGs_char, arrayType="450k", maxGap=100, minCpGs=3)

}
