% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindCloseCpGsRegions.R
\name{closeCpGs.regions}
\alias{closeCpGs.regions}
\title{Extract clusters of close CpGs}
\usage{
closeCpGs.regions(CpGs_char, arrayType = c("450k", "EPIC"),
  maxGap = 200, min.cpgs = 3)
}
\arguments{
\item{maxGap}{an integer, genomic locations within maxGap from each other are placed into the same cluster}

\item{min.cpgs}{an integer, minimum nubmer of cpgs for resulting regions}

\item{CpGs}{a list of CpG ids}

\item{location.file}{a dataframe with cpg name (cpg), chromosome (CHR) and location (MAPINFO), row.names = cpg}
}
\value{
a list, each item is a list of CpG ids in the contiguous region (i.e. same cluster)
}
\description{
Extract clusters of close CpGs
}
\examples{
   CpGsChr22_char<-c("cg02953382", "cg12419862", "cg24565820", "cg04234412",
      "cg04824771", "cg09033563", "cg10150615", "cg18538332", "cg20007245",
      "cg23131131", "cg25703541")

}
