% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoMethAllRegions.R
\name{CoMethAllRegions}
\alias{CoMethAllRegions}
\title{Extract contiguous comethylated CpG regions}
\usage{
CoMethAllRegions(closeByGenomicRegionFile, betaMatrix,
  arrayType = c("450k", "EPIC"), returnAllCpGs = TRUE,
  fileType = c("gmt", "RDS"), dataDir, ...)
}
\arguments{
\item{closeByGenomicRegionFile}{a file with the close by regions}

\item{betaMatrix}{matrix of beta values, with row names = CpG ids,
column names = sample ids}

\item{arrayType}{Type of array, 450k or EPIC}

\item{returnAllCpGs}{indicates if outputting all the CpGs in the region
when there is not a contiguous comethylated region or
only the CpGs in the contiguous comethylated regions}

\item{fileType}{the output files can be saved as .gmt or .RDS,
.gmt files can be open as flat text file, .RDS files are 50% the size of
.gmt files, but cannot be open}

\item{dataDir}{link to the directory where the filtered files will be saved}

\item{...}{}
}
\value{
a text file with a list of two items for each region:
   1. Contiguous_Regions - a data frame with CpG = CpG name,
   Chr = chromosome number,
   MAPINFO = genomic position,
   r_drop = correlation between the CpG with rest of the CpGs,
   keep = indicator for co-methylated CpG,
   keep_contiguous = cotiguous comethylated subregion number
   2. CpGs_subregions - lists of CpGs in each contiguous co-methylated
   subregion
   a file with the contiguous comethylated CpG regions
}
\description{
Extract contiguous comethylated CpG regions
}
