% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{coMethDMR_setup}
\alias{coMethDMR_setup}
\title{Cache sesameData at Package Load}
\arguments{
\item{libname}{path to package library}

\item{pkgname}{package name}
}
\description{
Check if the user has both the HM540 and EPIC manifests in their
  cache. The contents of the cache are checked via a call to the
  \code{ExperimentHub} function. If not all data components are available in
  the cache for these two platforms, we \code{\link[AnnotationHub]{query}} 
  the necessary data to save them to the cache for later use.
}
\details{
arguments are unused
}
\keyword{internal}
