% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/contiguous_regions_input_one_region_rdrop_LWtest.R
\name{contiguous.regions}
\alias{contiguous.regions}
\title{Extract contiguous co-methylated regions}
\usage{
contiguous.regions(beta.matrix, min.region.size, location.file, threshold.r)
}
\arguments{
\item{beta.matrix}{a matrix of beta values, with rownames = cpg ids, column names = sample ids}

\item{min.region.size}{min number of cpgs in the sub-regions}

\item{location.file}{a dataframe with cpg name (cpg), chromosome (CHR) and location (MAPINFO), row.names = cpg}

\item{threshold.r}{min correlation between a cpg in the region with the rest of the CpGs}
}
\value{
A list of three items:
1. n.sub-regions - number of contiguous co-methylated sub-regions
2. contiguous.regions - a data frame of probe location (ProbeID, CHR, MAPINFO),
   index (ind), r.drop (correlation between the cpg with rest of the CpGs),
   indicator for contiguous co-methylated CpGs (keep, keep.contiguous)
3. CpGs.subregions - lists of CpGs in each contiguous co-methylated regions
}
\description{
Extract contiguous co-methylated regions
}
\examples{



}
