% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CpGsInfoAllRegions.R
\name{CpGsInfoAllRegions}
\alias{CpGsInfoAllRegions}
\title{Test individual CpGs in more than one region}
\usage{
CpGsInfoAllRegions(AllRegionNames_char, betas_df, pheno_df, contPheno_char,
  covariates_char, arrayType = c("450k", "EPIC"))
}
\arguments{
\item{AllRegionNames_char}{vector of character strings with location info for one region in
this format: "chrxx:xxxxxx-xxxxxx"}

\item{betas_df}{data frame of beta values for all contiguous
comethylated regions, with row names = CpG ids, column names = sample ids}

\item{pheno_df}{a data frame with phenotype and covariates
(sample ID column = "Sample")}

\item{contPheno_char}{character string of the phenotype name}

\item{covariates_char}{character vector of covariate names}

\item{arrayType}{Type of array, 450k or EPIC}
}
\value{
a data frame with Region, cpg, chr, pos, slopeEstimate and slopePval for
   each CpG in the tested regions
}
\description{
Test individual CpGs in more than one region
}
\examples{
   data(betaMatrixChr22_df)
   data(pheno_df)
   AllRegionNames_char <- c("chr22:18267969-18268249", "chr22:18531243-18531447")

   CpGsInfoAllRegions(
     AllRegionNames_char,
     betas_df = betaMatrixChr22_df,
     pheno_df, contPheno_char = "stage",
     covariates_char = c("age.brain", "sex")
   )
}
