% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CpGsInfoOneRegion.R
\name{CpGsInfoOneRegion}
\alias{CpGsInfoOneRegion}
\title{Test individual CpGs}
\usage{
CpGsInfoOneRegion(regionName_char, betaMatrixAllRegions, pheno_df,
  contPheno_char, covariates_char, arrayType = c("450k", "EPIC"))
}
\arguments{
\item{regionName_char}{character string with location info for one region in
this format: "chrxx:xxxxxx-xxxxxx"}

\item{betaMatrixAllRegions}{matrix of beta values for all contiguous
comethylated regions, with row names = CpG ids, column names = sample ids}

\item{pheno_df}{a data frame with phenotype and covariates
(sample ID column = "Sample")}

\item{contPheno_char}{character string of the phenotype name}

\item{covariates_char}{character vector of covariate names}

\item{arrayType}{Type of array, 450k or EPIC}
}
\value{
a data frame with cpg, chr, pos, slopeEstimate and slopePval for
   each CpG in the region
}
\description{
Test individual CpGs
}
\examples{
   data(betaMatrixChr22_df)
   data(pheno_df)
   CpGsInfoOneRegion(regionName_char = "chr22:18267969-18268249",
    betaMatrixAllRegions = betaMatrixChr22_df, pheno_df, contPheno_char = "stage",
    covariates_char = c("age.brain", "sex"))
}
