% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateCpGsRegions.R
\name{CreateCpGsRegions}
\alias{CreateCpGsRegions}
\title{Create a list object with class CpGsRegions}
\usage{
CreateCpGsRegions(CpGs_ls)
}
\arguments{
\item{CpGs_ls}{a list where each item is a character vector of CpGs IDs in a
region. Each vector should be named with the region name.}
}
\value{
A list object with class CpGsRegions
}
\description{
Create a list object with class CpGsRegions
}
\examples{

   CpGsChr22_ls <- readRDS(
                      system.file ("extdata",
                                   "CpGislandsChr22_ex.RDS",
                                   package = 'coMethDMR',
                                   mustWork = TRUE
                                  )
   )
   CreateCpGsRegions(CpGsChr22_ls)

}
