% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateRdrop.R
\name{CreateRdrop}
\alias{CreateRdrop}
\title{Create R-drop values}
\usage{
CreateRdrop(data, method = c("pearson", "spearman"))
}
\arguments{
\item{data}{a dataframe with rownames = sample ids, column names = CpG ids.}

\item{method}{correlation method, can be pearson or spearman}
}
\value{
A data frame with the following columns:

\itemize{
  \item{\code{CpG} : }{CpG ID}
  \item{\code{r_drop} : }{The correlation between each CpG with the sum of
  the rest of the CpGs}
}
}
\description{
Create R-drop values
}
\examples{
   data(betaMatrix_ex1)
   CreateRdrop(data = betaMatrix_ex1, method = "pearson")

}
