% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_enhancer_probes.R
\docType{data}
\name{enhancer.probes}
\alias{enhancer.probes}
\title{Example of genic regions that contains the CpGs.}
\format{
A large character vector containing 102559 elements, 7.4 MB.
}
\source{

}
\usage{
enhancer.probes
}
\description{
Genic regions that contains the CpGs.
}
\keyword{datasets}
