% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindComethylatedRegions.R
\name{FindComethylatedRegions2}
\alias{FindComethylatedRegions2}
\title{Find contiguous comethylated regions}
\usage{
FindComethylatedRegions2(CpGs_df, minCpGs_int = 3)
}
\arguments{
\item{CpGs_df}{a dataframe with cpg name (CpG), keep=1/drop=0 (alpha), ind=1:number of CpGs in the region}

\item{minCpGs_int}{n integer, minimum nubmer of cpgs for resulting regions}
}
\value{
data frame with CpG and subregion number
}
\description{
Find contiguous comethylated regions
}
\examples{
attach("tests/betaCluster_mat_example1.RData")
CpGs_df <- MarkComethylatedCpGs(betaCluster_mat = betaMatrix_ex1)
FindComethylatedRegions2(CpGs_df)

}
