% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCpGsInRegion.R
\name{GetCpGsInRegion}
\alias{GetCpGsInRegion}
\title{Extract individual CpGs in a region}
\usage{
GetCpGsInRegion(regionName_char, arrayType = c("450k", "EPIC"))
}
\arguments{
\item{regionName_char}{character string with location info for one region in
this format: "chrxx:xxxxxx-xxxxxx"}

\item{arrayType}{Type of array, 450k or EPIC}
}
\value{
vector of CpGs in the region
}
\description{
Extract individual CpGs in a region
}
\examples{
   GetCpGsInRegion(
     regionName_char = "chr22:18267969-18268249",
     arrayType = "450k"
   )
}
