% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetResiduals.R
\name{GetResiduals}
\alias{GetResiduals}
\title{Get Residuals}
\usage{
GetResiduals(dnam_df, betaToM = TRUE, pheno_df, covariates_char)
}
\arguments{
\item{dnam_df}{data frame of beta values for all genomic regions,
with row names = CpG IDs, column names = sample IDs. This is often the
genome-wide array data.}

\item{betaToM}{indicates if converting methylation beta values to mvalues}

\item{pheno_df}{a data frame with phenotype and covariates, with variable
\code{Sample} indicating sample IDs.}

\item{covariates_char}{character vector for names of the covariate variables}
}
\value{
output a matrix of residual values same size as \code{dnam_df}
}
\description{
Get Residuals
}
\examples{
   data(betaMatrixChr22_df)
   data(pheno_df)

   GetResiduals(
     dnam_df = betaMatrixChr22_df[1:10, 1:10],
     betaToM = TRUE,
     pheno_df = pheno_df,
     covariates_char = c("age.brain", "sex", "Mplate")
   )

}
