% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ImportSesameData.R
\name{ImportSesameData}
\alias{ImportSesameData}
\title{Import Illumina manifests (sesameData versions)}
\usage{
ImportSesameData(manifest_char)
}
\arguments{
\item{manifest_char}{Which manifest should be loaded? Currently, this package
has been tested to work with 450k and EPIC arrays for HG19 and HG38.}
}
\description{
Load either the HM540 and EPIC manifests into working memory
}
\details{
This function assumes that the .onLoad() function has executed 
  properly and (therefore) that the necessary data sets are in the cache.
}
\examples{
  hm450k_gr <- ImportSesameData("HM450.hg19.manifest")
  head(hm450k_gr)

}
