% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmmTest.R
\name{lmmTest}
\alias{lmmTest}
\title{Fit Mixed Model}
\usage{
lmmTest(betaMatrix, pheno, method = c("randomCoefficientMixedModel",
  "mixedModelWithInterceptOnly"))
}
\arguments{
\item{betaMatrix}{matrix of beta values for one contiguous comethylated region,
with row names = CpG ids, column names = sample ids}

\item{pheno}{a data frame with phenotype and covariates}

\item{model}{model used to fit mixed model}
}
\value{
list of pvalue and median correlation
   for the contiguous comethylated region being tested
}
\description{
Fit Mixed Model
}
