% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmmTestAllRegions.R
\name{lmmTestAllRegions}
\alias{lmmTestAllRegions}
\title{Fit mixed model for all regions}
\usage{
lmmTestAllRegions(betaMatrixAllRegions, pheno_df, contPheno_char,
  covariates_char, inFile, outFile, inFileType = c("gmt", "RDS"),
  arrayType = c("450k", "EPIC"), returnAllCpGs = FALSE,
  modelType = c("randCoeffMixed", "mixed"), rDropThresh_num = 0.5)
}
\arguments{
\item{betaMatrixAllRegions}{matrix of beta values for all contiguous
comethylated regions, with row names = CpG ids, column names = sample ids}

\item{pheno_df}{a data frame with phenotype and covariates
(sample ID column = "Sample")}

\item{contPheno_char}{character string of the phenotype name}

\item{covariates_char}{character vector of covariate names}

\item{inFile}{name of input file with close by regions}

\item{outFile}{output csv file with the results for the mixed model analysis}

\item{inFileType}{extension of input file: "gmt" or "RDS"}

\item{arrayType}{Type of array, 450k or EPIC}

\item{returnAllCpGs}{indicates if outputting all the CpGs in the region
when there is not a contiguous comethylated region or
only the CpGs in the contiguous comethylated regions}

\item{modelType}{model used to fit mixed model}
}
\value{
text file with RegionID, p-value and median correlation value for
   each contiguous comethylated region tested.
}
\description{
Fit mixed model for all regions
}
\examples{
\dontrun{
  data(betaMatrixChr22_df)
   data(pheno_df)
   inFile <- system.file(
     "extdata", "CpGislandsChr22_ex.RDS",
      package = 'coMethDMR', mustWork = TRUE
   )

   lmmTestAllRegions(
     betaMatrixAllRegions = betaMatrixChr22_df,
     pheno_df,
     contPheno_char = "stage",
     covariates_char = c("age.brain", "sex"),
     inFile,
     outFile = "outEx.csv",
     inFileType = "RDS",
     arrayType = "450k",
     returnAllCpGs = FALSE,
     modelType = "randCoeffMixed",
     rDropThresh_num = 0.5
   )
}

}
