% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkComethylatedCpGs.R
\name{MarkComethylatedCpGs}
\alias{MarkComethylatedCpGs}
\title{Find contiguous comethylated regions}
\usage{
MarkComethylatedCpGs(betaCluster_mat, rDropThresh_num = 0.5)
}
\arguments{
\item{betaCluster_mat}{matrix of beta values, with rownames = sample ids,
column names = CpG ids, and ordered by genomic position}

\item{rDropThresh_num}{min correlation between a cpg in the region with the
rest of the CpGs}
}
\value{
dataframe of CpGs to keep or drop
}
\description{
Find contiguous comethylated regions
}
\examples{
data(betaCluster_mat_example1)
   MarkComethylatedCpGs(betaCluster_mat = betaMatrix_ex1)
   data(betaCluster_mat_example2)
   MarkComethylatedCpGs(betaCluster_mat = betaMatrix_ex2)
   data(betaCluster_mat_example3)
   MarkComethylatedCpGs(betaCluster_mat = betaMatrix_ex3)
}
