% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NameRegion2.R
\name{NameRegion2}
\alias{NameRegion2}
\title{Name a CpG region based on its genomic location}
\usage{
NameRegion2(CpGsOrdered_df)
}
\arguments{
\item{CpGsOrdered_df}{dataframe with cpg name as character (cpg),
chromosome number as integer (CHR) and genomic location as integer (MAPINFO), row.names = cpg}
}
\value{
genome location of the CpGs region
}
\description{
Name a CpG region based on its genomic location
}
