% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util1_OrderCpGsByLocation.R
\name{OrderCpGsByLocation}
\alias{OrderCpGsByLocation}
\title{Order CpGs by genomic location}
\usage{
OrderCpGsByLocation(CpGs_char, arrayType = c("450k", "EPIC"),
  output = c("vector", "dataframe"))
}
\arguments{
\item{CpGs_char}{vector of CpGs}

\item{arrayType}{Type of array, 450k or EPIC}

\item{output}{vector of CpGs or dataframe with CpGs, CHR, MAPINFO}
}
\value{
vector of ordered CpGs or dataframe with ordered CpGs, CHR, MAPINFO
}
\description{
Order CpGs by genomic location
}
\examples{
CpGs_char <- c("cg04677227", "cg07146435", "cg11632906", "cg20214853")
   OrderCpGsByLocation(CpGs_char, arrayType=c("EPIC"), output = "dataframe")

}
