% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_pheno_df.R
\docType{data}
\name{pheno_df}
\alias{pheno_df}
\title{Example phenotype data file from Prefrontal Cortex (PFC) Methylation Data of Alzheimer's Disease subjects}
\format{
A data frame containing variables for Braak stage (stage), subject.id, Batch (slide), Sex,
Sample, age of brain sample (age.brain)
}
\source{
GEO accession: GSE59685
}
\usage{
pheno_df
}
\description{
Subset of phenotype information for Alzheimer's methylation dataset.
}
\keyword{datasets}
