% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpgs_location_name.R
\name{region.name}
\alias{region.name}
\title{Title}
\usage{
region.name(cpgs, location.file)
}
\arguments{
\item{cpgs}{vector of CpGs}

\item{location.file}{a dataframe with cpg name (cpg), chromosome (CHR) and location (MAPINFO), row.names = cpg}
}
\value{
genome location of the CpGs region
}
\description{
Title
}
\examples{


}
