% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_regions.R
\docType{data}
\name{regions}
\alias{regions}
\title{Example of probes mapped to the genomic regions.}
\format{
A S4 object. A formal class GRanges. Class seqnames with 4 slots.
   Class ranges with 6 slots. Class strand with 4 slots.
   Class sequinfo with 4 slots. Class elementMetadata with 6 slots.
   Class strand with 4 slots.
}
\source{

}
\usage{
regions
}
\description{
Probes mapped to the genomic regions.
}
\keyword{datasets}
