% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util4_RegionsToRanges.R
\name{RegionsToRanges}
\alias{RegionsToRanges}
\title{Convert genomic regions in a data frame to GRanges format}
\usage{
RegionsToRanges(data_df)
}
\arguments{
\item{data_df}{dataframe with variable \code{regionName} as a character vector,
in this format: "chrxx:xxxxxx-xxxxxx"}
}
\value{
genomic regions in GRanges format
}
\description{
Convert genomic regions in a data frame to GRanges format
}
\examples{
 df <- data.frame ( regionName = c("chr22:19709548-19709755",
                                   "chr2:241721922-241722113"))

 RegionsToRanges (df)
}
