% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util3_SplitCpGsSubregionsDataFrameToList.R
\name{SplitCpGsSubregionsDataFrameToList}
\alias{SplitCpGsSubregionsDataFrameToList}
\title{Split a dataframe of CpGs and comethylated subregions to a list of CpGs in each subregion}
\usage{
SplitCpGsSubregionsDataFrameToList(CpGsSubregions_df, arrayType)
}
\arguments{
\item{CpGsSubregions_df}{data frame with CpG and subregion number}

\item{arrayType}{Type of array, 450k or EPIC}
}
\value{
a list of comethylated subregions CpGs for a pre-defined region
}
\description{
Split a dataframe of CpGs and comethylated subregions to a list of CpGs in each subregion
}
\examples{
data(betaCluster_mat_example4)
   CpGs_df <- MarkComethylatedCpGs(betaCluster_mat = betaMatrix_ex4)
   CpGsSubregions_df <- FindComethylatedRegions(CpGs_df)
   SplitCpGsSubregionsDataFrameToList(CpGsSubregions_df, "450k")

}
