% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WriteCloseByAllRegions.R
\name{WriteCloseByAllRegions}
\alias{WriteCloseByAllRegions}
\title{Extract clusters of close by CpGs from all genomic regions}
\usage{
WriteCloseByAllRegions(file, genomicRegionType = c("ISLAND", "TSS1500",
  "TSS200", "UTR5", "EXON1", "GENEBODY", "UTR3", "NSHORE", "SSHORE",
  "NSHELF", "SSHELF"), arrayType = c("450k", "EPIC"), maxGap = 200,
  minCpGs = 3, fileType = c("gmt", "RDS"), ...)
}
\arguments{
\item{file}{file where the close by regions will be saved}

\item{genomicRegionType}{Type of region}

\item{arrayType}{Type of array, 450k or EPIC}

\item{maxGap}{an integer, genomic locations within maxGap from each other
are placed into the same cluster}

\item{minCpGs}{an integer, minimum number of CpGs for resulting regions}

\item{fileType}{the output files can be saved as .gmt or .RDS,
.gmt files can be open as flat text file, .RDS files are 50% the size of
.gmt files, but they can only be read in R enviroment.}

\item{...}{}
}
\value{
a file with the close by regions
}
\description{
Extract clusters of close by CpGs from all genomic regions
}
\details{
For \code{maxGap} = 200 and \code{minCpGs} = 3, we have calculated
   the close-by regions output files already. They are in /inst/extdata/
   File extension should not be supplied. We add it automatically through
   the fileType argument.
   Creating and writing this file will take around 25 minutes for
   genomicRegionType = "ISLAND"
}
\examples{
\dontrun{
  CloseByAllRegions(
     genomicRegionType = "ISLAND", arrayType = "450k", maxGap = 50,
     minCpGs = 3, fileType = "gmt", file = "closeByRegions"
  )
}

}
