% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closeCpGs_regions_LWedit.R
\name{closeCpGs.regions}
\alias{closeCpGs.regions}
\title{Extract clusters of close CpGs}
\usage{
closeCpGs.regions(CpGs, location.file, maxGap, min.cpgs)
}
\arguments{
\item{CpGs}{a list of CpG ids}

\item{location.file}{a dataframe with cpg name (cpg), chromosome (CHR) and location (MAPINFO), row.names = cpg}

\item{maxGap}{an integer, genomic locations within maxGap from each other are placed into the same cluster}

\item{min.cpgs}{an integer, minimum nubmer of cpgs for resulting regions}
}
\value{
a list, each item is a list of CpG ids in the contiguous region (i.e. same cluster)
}
\description{
Extract clusters of close CpGs
}
\examples{

}
