% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoMethAllRegions.R
\name{CoMethAllRegions}
\alias{CoMethAllRegions}
\title{Extract contiguous co-methylated genomic regions from a list of
  pre-defined genomic regions}
\usage{
CoMethAllRegions(betaMatrix, betaToM = TRUE, method = c("pearson",
  "spearman"), rDropThresh_num = 0.4, arrayType = c("450k", "EPIC"),
  CpGs_ls = NULL, file = NULL, fileType = c("gmt", "RDS"),
  regionType = c("ISLAND", "NSHORE", "NSHELF", "SSHORE", "SSHELF",
  "TSS1500", "TSS200", "UTR5", "EXON1", "GENEBODY", "UTR3"),
  returnAllCpGs = FALSE, output = c("CpGs", "dataframe"),
  cluster = NULL, ...)
}
\arguments{
\item{betaMatrix}{matrix (or data frame) of beta values, with row names = CpG IDs,
column names = sample IDs. This is typically genome-wide methylation beta
values.}

\item{betaToM}{indicates if converting methylation beta values to mvalues}

\item{method}{method for computing correlation, can be "spearman" or "pearson"}

\item{rDropThresh_num}{thershold for min correlation between a cpg with sum
of the rest of the CpGs}

\item{arrayType}{Type of array, can be "450k" or "EPIC"}

\item{CpGs_ls}{list where each item is a character vector of CpGs IDs.}

\item{file}{an RDS or gmt file with clusters of CpG locations (i.e. CpGs
located closely to each other on the genome). This file can be generated
by the \code{\link{WriteCloseByAllRegions}} function.}

\item{fileType}{file extension for input file, can be "gmt" or "RDS"}

\item{regionType}{Type of input genomic regions (e.g. "ISLAND" for CpG island)}

\item{returnAllCpGs}{When there is not a contiguous comethylated region in
the inputing pre-defined region, \code{returnAllCpGs = 1} indicates
outputting all the CpGs in the input regions, while
\code{returnAllCpGs = 0} indicates not returning any CpG.}

\item{output}{a character vector of CpGs or a dataframe of CpGs along with
rDrop info}

\item{...}{Dots for internal arguments. Currently unused.}
}
\value{
When \code{output = "dataframe"} is selected, returns a list of data frames, each with \code{CpG}
(CpG name), \code{Chr} (chromosome number), \code{MAPINFO} (genomic
position), \code{r_drop} (correlation between the CpG with rest of the
CpGs), \code{keep} (indicator for co-methylated CpG),
\code{keep_contiguous} (index for contiguous comethylated subregions).

When \code{output = "CpGs"} is selected, returns a list, each item is a list of CpGs
in the contiguous co-methylated subregion.
}
\description{
Extract contiguous co-methylated genomic regions from a list of
  pre-defined genomic regions
}
\details{
There are several ways to input genomic regions for this function: (1) use \code{CpGs_ls}
argument (2) use \code{regionType} argument (3) use \code{file} and \code{fileType} arguments,
examples of these files are at https://github.com/lissettegomez/coMethDMRdata
}
\examples{
   data(betaMatrixChr22_df)


   CpGisland_ls <- readRDS(
     system.file(
       "extdata",
       "CpGislandsChr22_ex.RDS",
       package = 'coMethDMR',
       mustWork = TRUE
     )
   )

   coMeth_ls <- CoMethAllRegions (
     betaMatrix = betaMatrixChr22_df,
     betaToM = TRUE,
     method = "pearson",
     CpGs_ls = CpGisland_ls,
     arrayType = "450k",
     returnAllCpGs = FALSE,
     output = "CpGs"
   )

\dontrun{

 CoMethAllRegions (
   betaMatrix = betaMatrixChr22_df,
   regionType = "ISLAND",
   arrayType = "450k",
   returnAllCpGs = FALSE
 )

}

}
