% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoMethAllRegions.R
\name{CoMethAllRegions}
\alias{CoMethAllRegions}
\title{Extract contiguous co-methylated genomic regions from a list of
  pre-defined genomic regions}
\usage{
CoMethAllRegions(betaMatrix, betaToM = TRUE, regionType = c("ISLAND",
  "NSHORE", "NSHELF", "SSHORE", "SSHELF", "TSS1500", "TSS200", "UTR5",
  "EXON1", "GENEBODY", "UTR3"), arrayType = c("450k", "EPIC"),
  file = NULL, fileType = c("gmt", "RDS"), CpGs_ls = NULL,
  rDropThresh_num = 0.4, method = c("pearson", "spearman"),
  returnAllCpGs = FALSE, output = c("CpGs", "dataframe"), ...)
}
\arguments{
\item{betaMatrix}{matrix (or data frame) of beta values, with row names = CpG IDs,
column names = sample IDs. This is typically genome-wide methylation beta
values.}

\item{regionType}{Type of input genomic regions (e.g. "ISLAND" for CpG island)}

\item{arrayType}{Type of array, can be "450k" or "EPIC"}

\item{file}{an RDS or gmt file with clusters of CpG locations (i.e. CpGs
located closely to each other on the genome). This file can be generated
by the \code{\link{WriteCloseByAllRegions}} function.}

\item{fileType}{file extension for input file, can be "gmt" or "RDS"}

\item{CpGs_ls}{list where each item is a character vector of CpGs IDs.}

\item{rDropThresh_num}{thershold for min correlation between a cpg with sum
of the rest of the CpGs}

\item{returnAllCpGs}{When there is not a contiguous comethylated region in
the inputing pre-defined region, \code{returnAllCpGs = 1} indicates
outputting all the CpGs in the input regions, while
\code{returnAllCpGs = 0} indicates not returning any CpG.}

\item{...}{Dots for internal arguments. Currently unused.}
}
\value{
A list of two components:
  \itemize{
    \item{\code{Contiguous_Regions} : }{A data frame with \code{CpG}
      (CpG name), \code{Chr} (chromosome number), \code{MAPINFO} (genomic
      position), \code{r_drop}(correlation between the CpG with rest of the
      CpGs), \code{keep} (indicator for co-methylated CpG),
      \code{keep_contiguous} (index for contiguous comethylated subregions)
    }
    \item{\code{CpGsSubregions} : } {results from all the regions, each item
      is a list of CpGs in the contiguous co-methylated subregion
    }
  }
}
\description{
Extract contiguous co-methylated genomic regions from a list of
  pre-defined genomic regions
}
\examples{


   data(betaMatrixChr22_df)

   CpGsChr22_ls <- readRDS(
                      system.file ("extdata",
                                   "CpGislandsChr22_ex.RDS",
                                   package = 'coMethDMR',
                                   mustWork = TRUE
                                  )
   )


   CoMethAllRegions (
     betaMatrix = betaMatrixChr22_df,
     CpGs_ls = CpGsChr22_ls,
     arrayType = "450k",
     returnAllCpGs = FALSE
   )




\dontrun{

CoMethAllRegions (
     betaMatrix = betaMatrixChr22_df,
     regionType = "ISLAND",
     arrayType = "450k",
     returnAllCpGs = FALSE
   )

}

}
