% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoMethAllRegions.R
\name{CoMethAllRegions}
\alias{CoMethAllRegions}
\title{Extract contiguous comethylated CpG regions}
\usage{
CoMethAllRegions(betaMatrix, file, fileType = c("gmt", "RDS"),
  arrayType = c("450k", "EPIC"), returnAllCpGs = FALSE, ...)
}
\arguments{
\item{betaMatrix}{matrix of beta values, with row names = CpG ids,
column names = sample ids}

\item{file}{name of the file with the close by regions}

\item{fileType}{the output files can be saved as .gmt or .RDS,
.gmt files can be open as flat text file, .RDS files are 50% the size of
.gmt files, but cannot be open}

\item{arrayType}{Type of array, 450k or EPIC}

\item{returnAllCpGs}{indicates if outputting all the CpGs in the region
when there is not a contiguous comethylated region or
only the CpGs in the contiguous comethylated regions}

\item{...}{}
}
\value{
a list of two items:
   1. Contiguous_Regions - a data frame with
   CpG = CpG name,
   Chr = chromosome number,
   MAPINFO = genomic position,
   r_drop = correlation between the CpG with rest of the CpGs,
   keep = indicator for co-methylated CpG,
   keep_contiguous = cotiguous comethylated subregion number
   for all the regions
   2. CpGs_subregions - lists of CpGs in each contiguous co-methylated
   subregion from all the regions
}
\description{
Extract contiguous comethylated CpG regions
}
\examples{
   data(betaMatrixChr22_df)
   CoMethAllRegions(
      betaMatrix = betaMatrixChr22_df,
      file = system.file(
         "extdata",
         "CpGislandsChr22_ex.RDS",
         package = 'coMethDMR',
         mustWork = TRUE),
      fileType = "RDS",
      arrayType = "450k",
      returnAllCpGs = FALSE
      )

}
