% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coMethWrapper.R
\name{coMethWrapper}
\alias{coMethWrapper}
\title{Wrapper Function to find comethyalted regions}
\usage{
coMethWrapper(CpGs_char, betaMatrix)
}
\arguments{
\item{CpGs_char}{vector of CpGs}

\item{betaMatrix}{matrix of beta values, with rownames = CpG ids,
column names = sample ids}
}
\value{
list of two items:
  1. Contiguous_Regions - a data frame of CpG location (CpG, Chr, MAPINFO),
}
\description{
Wrapper Function to find comethyalted regions
}
\examples{
   data(betaMatrixChr22_df)
   CpGsChr22_char<-c("cg02953382", "cg12419862", "cg24565820", "cg04234412",
      "cg04824771", "cg09033563", "cg10150615", "cg18538332", "cg20007245",
      "cg23131131", "cg25703541")
   coMethWrapper(CpGsChr22_char, betaMatrixChr22_df)
}
