% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coMethWrapper.R
\name{coMethWrapper}
\alias{coMethWrapper}
\title{Wrapper Function to find comethyalted regions}
\usage{
coMethWrapper(CpGs_char, betaMatrix)
}
\arguments{
\item{CpGs_char}{vector of CpGs}

\item{betaMatrix}{matrix of beta values, with rownames = CpG ids,
column names = sample ids}
}
\value{
list of two items:
  1. Contiguous_Regions - a data frame of CpG location (CpG, Chr, MAPINFO),
}
\description{
Wrapper Function to find comethyalted regions
}
\examples{
   beta <- read.csv("C:/Users/lxg255/Box Sync/Rpackage-coMethDMR/PFCbetas_BMIQnormalized.csv")
   row.names(beta) <- beta[,1]
   betaMatrix <- beta[,-1]
   data(betaCluster_mat_example4)
   CpGs_char <- colnames(betaMatrix_ex4)
   coMethWrapper(CpGs_char, betaMatrix)
}
