% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CpGsInfoOneRegion.R
\name{CpGsInfoOneRegion}
\alias{CpGsInfoOneRegion}
\title{Test associations of individual CpGs in a genomic region with a continuous phenotype}
\usage{
CpGsInfoOneRegion(regionName_char, betas_df, pheno_df, contPheno_char,
  covariates_char, arrayType = c("450k", "EPIC"))
}
\arguments{
\item{regionName_char}{character string of location information for a genomic region, specified in
the format of "chrxx:xxxxxx-xxxxxx"}

\item{betas_df}{data frame of beta values for all contiguous
comethylated regions, with row names = CpG IDs, column names = sample IDs}

\item{pheno_df}{a data frame with phenotype and covariate variables, with variable "Sample" for sample IDs.}

\item{contPheno_char}{character string of the continuous phenotype, to be tested against methylation values}

\item{covariates_char}{character vector of covariate variables names}

\item{arrayType}{Type of array, can be "450k" or "EPIC"}
}
\value{
a data frame with CpG ID (cpg), chromosome (chr), position (pos), and results for testing
association of methylation in individual CpGs with continuous phenotype (slopeEstimate, slopePval)
}
\description{
Test associations of individual CpGs in a genomic region with a continuous phenotype
}
\examples{
   data(betaMatrixChr22_df)
   data(pheno_df)

   CpGsInfoOneRegion(
     regionName_char = "chr22:18267969-18268249",
     betas_df = betaMatrixChr22_df,
     pheno_df, contPheno_char = "stage",
     covariates_char = c("age.brain", "sex")
   )
}
