% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhancer_probes.R
\docType{data}
\name{enhancer.probes}
\alias{enhancer.probes}
\title{Enhancer HM450 probes}
\format{
A list of HM450 probes names
}
\source{
Illumina array
}
\usage{
enhancer.probes
}
\description{
list of HM450 probes names defined as enhancers
}
\keyword{datasets}
