% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindComethylatedRegions.R
\name{FindComethylatedRegions}
\alias{FindComethylatedRegions}
\title{Find contiguous comethylated regions}
\usage{
FindComethylatedRegions(betaCluster_mtx, minCpGs_int = 3,
  CpGlocations_df, threshold_r_num = 0.5)
}
\arguments{
\item{betaCluster_mtx}{matrix of beta values}

\item{minCpGs_int}{n integer, minimum nubmer of cpgs for resulting regions}

\item{CpGlocations_df}{a dataframe with cpg name (cpg), chromosome (CHR) and location (MAPINFO), row.names = cpg}

\item{threshold_r_num}{min correlation between a cpg in the region with the rest of the CpGs}
}
\description{
Find contiguous comethylated regions
}
\examples{

}
