% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCpGsInRegion.R
\name{GetCpGsInRegion}
\alias{GetCpGsInRegion}
\title{Extract probe IDs for CpGs located in a genomic region}
\usage{
GetCpGsInRegion(
  regionName_char,
  genome = c("hg19", "hg38"),
  arrayType = c("450k", "EPIC")
)
}
\arguments{
\item{regionName_char}{character string with location information for one
region in the format \code{"chrxx:xxxxxx-xxxxxx"}}

\item{genome}{human genome of reference hg19 (default) or hg38}

\item{arrayType}{Type of array, 450k or EPIC}
}
\value{
vector of CpG probe IDs mapped to the genomic region
}
\description{
Extract probe IDs for CpGs located in a genomic region
}
\examples{
   GetCpGsInRegion(
     regionName_char = "chr22:18267969-18268249",
     genome = "hg19",
     arrayType = "450k"
   )
}
