% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetResiduals.R
\name{GetResiduals}
\alias{GetResiduals}
\title{Get Residuals}
\usage{
GetResiduals(dnam, betaToM = TRUE, pheno_df, covariates_char)
}
\arguments{
\item{dnam}{data frame or matrix of methylation values,
with row names = CpG IDs, column names = sample IDs. This is often the
genome-wide array data. Note that if beta values are the input here,
then \code{betaToM} should be set to \code{TRUE}.
If mvalues are the input, then \code{betaToM} should be set to \code{FALSE}}

\item{betaToM}{indicates if converting methylation beta values to mvalues}

\item{pheno_df}{a data frame with phenotype and covariates, with variable
\code{Sample} indicating sample IDs.}

\item{covariates_char}{character vector for names of the covariate variables}
}
\value{
output a matrix of residual values, in the same dimension as \code{dnam}
}
\description{
Get Residuals
}
\examples{
   data(betaMatrixChr22_df)

   data(pheno_df)

   GetResiduals(
     dnam = betaMatrixChr22_df[1:10, 1:10],
     betaToM = TRUE,
     pheno_df = pheno_df,
     covariates_char = c("age.brain", "sex", "Mplate")
   )

}
