% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetResiduals.R
\name{GetResiduals}
\alias{GetResiduals}
\title{Get Residuals}
\usage{
GetResiduals(beta_df, pheno_df, covariates_char, na.action = na.exclude)
}
\arguments{
\item{beta_df}{data frame of beta values for all genomic regions,
with row names = CpG IDs, column names = sample IDs. This is often the
genome-wide array data.}

\item{pheno_df}{a data frame with phenotype and covariates, with variable
\code{Sample} indicating sample IDs.}

\item{covariates_char}{character vector for names of the covariate variables}

\item{na.action}{see details in na.action argument in function lm}
}
\value{
output a matrix of residual values same size as \code{beta_mat}
}
\description{
Get Residuals
}
\examples{
   data(betaMatrixChr22_df)
   data(pheno_df)

   GetResiduals(
     beta_df = betaMatrixChr22_df[1:10, 1:10],
     pheno_df = pheno_df,
     covariates_char = c("age.brain", "sex", "Mplate")
   )

}
