% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmmTest.R
\name{lmmTest}
\alias{lmmTest}
\title{Fit mixed model for one region}
\usage{
lmmTest(betaMatrix, pheno_df, contPheno_char, covariates_char,
  modelType = c("randCoeffMixed", "mixed"), arrayType = c("450k",
  "EPIC"))
}
\arguments{
\item{betaMatrix}{matrix of beta values for one contiguous comethylated region,
with row names = CpG ids, column names = sample ids}

\item{pheno_df}{a data frame with phenotype and covariates
(sample ID column = "Sample")}

\item{contPheno_char}{character string of the phenotype name}

\item{covariates_char}{character vector of covariate names}

\item{modelType}{model used to fit mixed model}
}
\value{
list of pvalue and median correlation
   for the contiguous comethylated region being tested
}
\description{
Fit mixed model for one region
}
\examples{
  data(betaMatrixChr22_df)
  CpGsChr22_char<-c("cg02953382", "cg12419862", "cg24565820", "cg04234412",
      "cg04824771", "cg09033563", "cg10150615", "cg18538332", "cg20007245",
      "cg23131131", "cg25703541")
  coMethCpGs <- CoMethSingleRegion(CpGsChr22_char, betaMatrixChr22_df)
  coMethBetaMatrix <- betaMatrixChr22_df[coMethCpGs$CpGsSubregions[[1]], ]
  data(pheno_df)
  lmmTest(betaMatrix = coMethBetaMatrix, pheno_df, contPheno_char = "stage",
      covariates_char = c("age.brain", "sex"), modelType = "randCoeffMixed",
      arrayType = "450k")
}
