% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmmTestAllRegions.R
\name{lmmTestAllRegions}
\alias{lmmTestAllRegions}
\title{Fit mixed model to test association between a continuous phenotype and
methylation values in a list of genomic regions}
\usage{
lmmTestAllRegions(beta_df, region_ls, pheno_df, contPheno_char,
  covariates_char, modelType = c("randCoef", "simple"),
  arrayType = c("450k", "EPIC"), outFile = NULL)
}
\arguments{
\item{beta_df}{data frame of beta values for all genomic regions,
with row names = CpG IDs, column names = sample IDs. This is often the
genome-wide array data.}

\item{region_ls}{a list of genomic regions, each item is a vector of CpG IDs within a genomic region. The co-methylated
regions can be obtained by function \code{CoMethAllRegions}.}

\item{pheno_df}{a data frame with phenotype and covariates, with variable
\code{Sample} indicating sample IDs.}

\item{contPheno_char}{character string of the main effect (a continuous
phenotype) to be tested for association with methylation values in each
region}

\item{covariates_char}{character vector for names of the covariate variables}

\item{modelType}{type of mixed model, can be \code{randCoef} for random
coefficient mixed model, or \code{simple} for simple linear mixed model.}

\item{arrayType}{Type of array, can be "450k" or "EPIC"}

\item{outFile}{output .csv file with the results for the mixed model analysis}
}
\value{
csv file with location of the genomic region (\code{chrom, start, end}), number of CpGs (\code{nCpGs}),
\code{Estimate}, Standard error (\code{StdErr}) of the test statistic, p-value and False Discovery Rate (FDR)
for association between methylation values in each genomic region with phenotype (\code{pValue}).
}
\description{
Fit mixed model to test association between a continuous phenotype and
methylation values in a list of genomic regions
}
\examples{
   data(betaMatrixChr22_df)

   data(pheno_df)

   CpGisland_ls <- system.file(
     "extdata", "CpGislandsChr22_ex.RDS",
      package = 'coMethDMR', mustWork = TRUE
   )

   coMeth_ls <- CoMethAllRegions(
                   betaMatrix = betaMatrixChr22_df,
                   file = CpGisland_ls,
                   fileType = "RDS",
                   arrayType = "450k",
                   rDropThresh_num = 0.4,
                   returnAllCpGs = FALSE
               )

   lmmTestAllRegions(
     beta_df = betaMatrixChr22_df,
     region_ls = coMeth_ls$CpGsSubregions,
     pheno_df,
     contPheno_char = "stage",
     covariates_char = c("age.brain", "sex"),
     modelType = "randCoef",
     arrayType = "450k"
   )

}
