% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NameRegion.R
\name{NameRegion}
\alias{NameRegion}
\title{Name a CpG region based on its genomic location}
\usage{
NameRegion(CpGs_char, CpGlocations_df = CpGlocations_df)
}
\arguments{
\item{CpGs_char}{vector of CpGs}

\item{CpGlocations_df}{dataframe with cpg name as character (cpg),
chromosome number as integer (CHR) and genomic location as integer (MAPINFO), row.names = cpg}
}
\value{
genome location of the CpGs region
}
\description{
Name a CpG region based on its genomic location
}
\examples{
CpGs_char <- c("cg04677227", "cg07146435", "cg11632906", "cg20214853")
NameRegion(CpGs_char, CpGlocations_df)

}
