% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util2_NameRegion.R
\name{NameRegion}
\alias{NameRegion}
\title{Name a region with several CpGs based on its genomic location}
\usage{
NameRegion(CpGsOrdered_df)
}
\arguments{
\item{CpGsOrdered_df}{dataframe with columns for Probe IDs as character (cpg),
chromosome number as character (chr) and genomic location as integer (pos)}
}
\value{
genome location of the CpGs, in the format of "chrxx:xxxxxx-xxxxxx"
}
\description{
Name a region with several CpGs based on its genomic location
}
\examples{
  data(CpGsOrdered_df)
  NameRegion(CpGsOrdered_df)

}
