% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util2_NameRegion.R
\name{NameRegion}
\alias{NameRegion}
\title{Name a region with several CpGs based on its genomic location}
\usage{
NameRegion(CpGsOrdered_df)
}
\arguments{
\item{CpGsOrdered_df}{dataframe with columns for Probe IDs as character (cpg),
chromosome number as character (chr) and genomic location as integer (pos)}
}
\value{
genome location of the CpGs, in the format of "chrxx:xxxxxx-xxxxxx"
}
\description{
Name a region with several CpGs based on its genomic location
}
\examples{

 CpGsOrdered_df <- data.frame(
  "chr" = c("chr10","chr10"),
  "pos" = c(100028236,100028499)
 )
 NameRegion(CpGsOrdered_df)

\dontrun{
  CpGs_char <- c("cg04677227", "cg07146435", "cg11632906", "cg20214853")
  CpGsOrdered_df <- OrderCpGsByLocation(CpGs_char, arrayType=c("EPIC"), output = "dataframe")
  NameRegion(CpGsOrdered_df)
}
}
