% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OrderCpGsByLocation.R
\name{OrderCpGsByLocation}
\alias{OrderCpGsByLocation}
\title{Order CpGs by genomic location}
\usage{
OrderCpGsByLocation(CpGs_char, CpGlocations_df, Output = c("CpGs",
  "CpGsOrdered_df"))
}
\arguments{
\item{CpGs_char}{vector of CpGs}

\item{CpGlocations_df}{dataframe with cpg name as character (cpg),
chromosome number as integer (CHR) and genomic location as integer (MAPINFO), row.names = cpg}

\item{vector}{Output: vector of CpGs or dataframe with CpGs, CHR, MAPINFO}
}
\value{
of ordered CpGs or dataframe of ordered CpGs, CHR, MAPINFO
}
\description{
Order CpGs by genomic location
}
\examples{

}
