% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util4_RegionsToRanges.R
\name{RegionsToRanges}
\alias{RegionsToRanges}
\title{Convert genomic regions in a data frame to GRanges format}
\usage{
RegionsToRanges(regionName_char)
}
\arguments{
\item{regionName_char}{a character vector of regions in the format
\code{"chrxx:xxxxxx-xxxxxx"}}
}
\value{
genomic regions in GRanges format
}
\description{
Convert genomic regions in a data frame to GRanges format
}
\examples{
 regions <- c("chr22:19709548-19709755", "chr2:241721922-241722113")
 RegionsToRanges(regions)
}
