% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util3_SplitCpGsSubregionsDataFrameToList.R
\name{SplitCpGDFbyRegion}
\alias{SplitCpGDFbyRegion}
\title{Split CpG dataframe by Subregion}
\usage{
SplitCpGDFbyRegion(CpGsSubregions_df, arrayType = c("450k", "EPIC"),
  returnAllCpGs = TRUE)
}
\arguments{
\item{CpGsSubregions_df}{data frame with CpG and subregion number}

\item{arrayType}{Type of array, 450k or EPIC}

\item{returnAllCpGs}{indicates if outputting all the CpGs in the region
when there is not a contiguous comethylated region or
only the CpGs in the contiguous comethylated regions}
}
\value{
a list of comethylated subregions CpGs for a pre-defined region
}
\description{
Split a dataframe of CpGs and comethylated subregions to a list
   of CpGs in each subregion
}
\examples{
   data(betaCluster_mat_example4)
   CpGs_df <- MarkComethylatedCpGs(betaCluster_mat = betaMatrix_ex4)
   CpGsSubregions_df <- FindComethylatedRegions(CpGs_df)
   SplitCpGDFbyRegion(CpGsSubregions_df, "450k")

}
\keyword{internal}
