% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WriteCloseByAllRegions.R
\name{WriteCloseByAllRegions}
\alias{WriteCloseByAllRegions}
\title{Extract clusters of CpG probes located closely}
\usage{
WriteCloseByAllRegions(
  fileName,
  regions,
  genome = c("hg19", "hg38"),
  arrayType = c("450k", "EPIC"),
  maxGap = 200,
  minCpGs = 3,
  ...
)
}
\arguments{
\item{fileName}{Name of the RDS file where the output genomic regions will be saved.}

\item{regions}{GRanges of input genomic regions}

\item{genome}{Human genome of reference hg19 or hg38}

\item{arrayType}{Type of array, can be "450k" or "EPIC"}

\item{maxGap}{an integer, genomic locations within maxGap from each other
are placed into the same cluster}

\item{minCpGs}{an integer, minimum number of CpGs for each resulting region}

\item{...}{Dots for internal arguments. Currently unused.}
}
\value{
a file with the genomic regions containing CpGs located closely within each
 inputing pre-defined genomic region
}
\description{
Extract clusters of CpG probes located closely
}
\details{
For \code{maxGap} = 200 and \code{minCpGs} = 3, we already calculated
   the clusters of CpGs. They are saved in folder \code{/inst/extdata/}.
}
\examples{

data(regions)

dontrun{
  WriteCloseByAllRegions(
    regions = regions, arrayType = "EPIC", maxGap = 50,
    minCpGs = 3, fileName = "closeByRegions.rds"
  )
}

}
